+++ contracts/ForAndWhileLoops-13.sol
@@ -4,28 +4,28 @@ pragma solidity ^0.8.7;
 
 contract ForAndWhileLoops {
     function loops() external pure {
-        for(uint i = 0; i < 10; i++) {
-            if(i < 3) {
+        for (uint256 i = 0; i < 10; i++) {
+            if (i < 3) {
                 continue;
             }
 
-            if(i == 5) {
+            if (i == 5) {
                 break;
             }
         }
 
-        uint j = 0;
-        while(j < 3) {
+        uint256 j = 0;
+        while (j < 3) {
             j++;
         }
     }
 
-    function sum(uint _n) external pure returns (uint) {
+    function sum(uint256 _n) external pure returns (uint256) {
         // 222 - 106112 gas
         // 2 - 22952 gas
         // 100 - 59996 gas
-        uint s;
-        for(uint i = 0; i <= _n; i++) {
+        uint256 s;
+        for (uint256 i = 0; i <= _n; i++) {
             s += i;
         }
         return s;
