+++ contracts/Mapping-22.sol
@@ -7,15 +7,15 @@ pragma solidity ^0.8.7;
 // Set, get, delete
 
 contract Mapping {
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
     mapping(address => mapping(address => bool)) public isFriend;
 
     function examples() external {
         balances[msg.sender] = 123;
 
-        uint bal = balances[msg.sender];
+        uint256 bal = balances[msg.sender];
 
-        uint bal3 =  balances[address(1)]; // 0
+        uint256 bal3 = balances[address(1)]; // 0
 
         balances[msg.sender] += 223; // 123 + 223 = 346
